pipes now work directly
